+++ src/main/java/me/jumper251/replay/replaysystem/data/types/VelocityData.java
@@ -7,15 +7,15 @@ public class VelocityData extends PacketData {
 	 */
 	private static final long serialVersionUID = 4046621273672598227L;
 	
-	private double x, y, z;
+	private float x, y, z;
 	
 	private int id;
 	
 	public VelocityData(int id, double x, double y, double z) {
 		this.id = id;
-		this.y = y;
-		this.x = x;
-		this.z = z;
+		this.y = (float) y;
+		this.x = (float) x;
+		this.z = (float) z;
 	}
 	
 	public int getId() {
